refactor(cli): move list sub-app commands to app/cli/list.py
Moves list_conduits_cmd and list_flows_cmd. list_app itself remains in
cli.main; this module just decorates it. Drops orphaned re-imports
(Counter, Table, Progress, parse_flow_id, _flow_duration_seconds,
_format_clock, _format_duration_seconds, _FLOW_STATUS_STYLE,
_task_status_summary) from app/main.py — those names had no remaining
users. Keeps _render_task_event and _truncate_tail re-exports until
Task 12 updates tests/test_main.py to canonical imports.
